utils/new-device-stats: add script from Martijn
This is a nice script that generats the "xx new devices" section for
blog posts, optionally with a chart (see one year post).

Original source from:
https://github.com/MartijnBraam/pmos-stats

Adding it here, so I can make a MR with changes I did for the latest
post on top, and so it is on the pmOS GitLab project.