index: update quotes (!83)
Remove dates and parts of quotes that would become dated soon, so we
don't need to put effort into updating the quotes. (Having old dates on
quotes makes the project seem dead on the front page, although it's very
much alive and kicking in the git repos, issues etc.). Replace Drew's
quote with a new one and reorder the quotes. Add a short description of
the author of each quote.