CSS tweaks for mobile view
Make the disclaimer font size smaller, so the "yet." is not in a new
line, and so it has the same font-size as the "latest-news" class.

On the start page, make the padding around the slogan smaller, the total
height of everything on top of the header image is not higher than the
background image's height. This fixes a big weird white border, which
appeared on top of the background image.